application/requirements.txt
fastapi==0.85.0
pydantic==1.9.2
uvicorn==0.18.3
python-dotenv
pyyaml
loguru==0.6.0